# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Workyy Realtime API
  version: 0.2.0
servers:
  - url: https://api.workyy.dev
    description: Production
  - url: http://localhost:4000/api
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Problem:
      type: object
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
          example: https://api.workyy.dev/problems/conflict
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: object
          additionalProperties: true
...
paths:
...
  /secrets:
    get:
      summary: List workspace secrets
      operationId: listSecrets
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: workspaceId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Secret list
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, workspaceId, name, createdAt, updatedAt]
                      properties:
                        id: { type: string, format: uuid }
                        workspaceId: { type: string, format: uuid }
                        name: { type: string }
                        createdAt: { type: string, format: date-time }
                        updatedAt: { type: string, format: date-time }
        '422':
          description: Validation error
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
    post:
      summary: Store a workspace secret
      operationId: createSecret
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId, name, value]
              properties:
                workspaceId:
                  type: string
                  format: uuid
                name:
                  type: string
                value:
                  type: string
      responses:
        '201':
          description: Secret stored
          content:
            application/json:
              schema:
                type: object
                required: [id, workspaceId, name, createdAt]
                properties:
                  id: { type: string, format: uuid }
                  workspaceId: { type: string, format: uuid }
                  name: { type: string }
                  createdAt: { type: string, format: date-time }
        '404':
          description: Workspace not found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: Secret name conflict
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          description: Validation error
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }